requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cmake==3.17.2
cycler==0.10.0
decorator==4.4.2
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.9
imagecodecs==2020.2.18

imageio==2.8.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.15.1
kiwisolver==1.2.0
matplotlib
networkx==2.4
numpy==1.21.2
opencv-python-headless==4.2.0.34
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy
six==1.14.0
threadpoolctl==2.0.0
tifffile==2020.5.11
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
